Removed problematic methods; rails3 seems to work
Rails has changed the way callbacks work; now the public-facing
api is much better, and can be used outside ActiveRecord, rails, etc.
Unfortunately, it looks like it might be difficult to make the way
attachment_fu hooks into rails 2.X work in rails 3. The documentation
for set_callback alludes to something that might work (specifically,
supporting passing in a non-self argument to the callback), but
the examples it uses in this regard don't seem to be accurate anymore.

Oh well, this *seems* to work for now.